refactor(DatasetExplorer): extract types and hooks (#119)
Phase 1 of the DatasetExplorer refactoring epic (#92). Moves the
monolithic 6,175-line component into a directory module and extracts
shared types, constants, and three custom hooks:

- types.ts: 14 interfaces, 8 constants, 2 utility functions
- useFilterPresets: preset CRUD, localStorage persistence
- useViewPreferences: chart/table toggle, percentage labels, survival
  view, settings menu click-outside
- useDashboard: dashboard chart list, saved dashboards API CRUD,
  IntersectionObserver for lazy loading

DatasetExplorer.tsx: 6,175 → 5,489 lines (-686).
Adds 25 hook unit tests (130 total client tests, all passing).
