+++ client/src/pages/DatasetExplorer/index.tsx
@@ -1,0 +1,2 @@
+export { default } from './DatasetExplorer'
+export { persistChartOverrides, loadChartOverrides } from './types'
